Return the number of elements in a linked list

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -9,12 +9,12 @@
 
 size_t listint_len(const listint_t *h)
 {
-	unsigned int count;
+	size_t count;
 
 	if (h == NULL)
 		return (0);
 
-	for (count = 0; h != NULL; count ++)
+	for (count = 0; h != NULL; count++)
 	{
 		h = h->next;
 	}
